[ADD] Rook ChessPiece 로직 추가 및 calculateUnitDirection 메소드를 RelativeChessPoint로 이동
