+++ src/main/java/chess/domain/chesspiece/Queen.java
@@ -20,10 +20,4 @@ public class Queen implements ChessPiece {
 
         return DIRECTIONS.stream().anyMatch(d -> d.equals(unitDirection));
     }
-
-    @Override
-    public RelativeChessPoint calculateUnitDirection(ChessPoint source, ChessPoint target) {
-        RelativeChessPoint relativePoint = target.minus(source);
-        return relativePoint.toUnit();
-    }
 }
